requirements.txt
streamlit>=1.28.0
annotated-types
pydantic
pydantic-core
typing-inspection
openai>=1.0.0
python-dotenv